docs
A simple api for a blog using a generic class for presentation. And also by user token authentication using the built-in django rest framework tools.  Customized custom permissions to access articles by role on the site.

To run this api you will need to install in a viral environment:
'django', 'djangorestframework', 'django-filte'